native-longlong: merged 1839-1883 from trunk
git-svn-id: svn://localhost/public/MoSync/branches/native-longlong@1885 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ tools/debugger/cmd_break.cpp
@@ -46,7 +46,7 @@ void break_enable(const string& args);
 InstructionMap sInstructions;
 BreakpointMap sBreakpoints;
 BreakpointAddressMap sBreakpointAddresses;
-TempBreakpoint gTempBreakpoint = { NULL, -1, 0 };
+TempBreakpoint gTempBreakpoint = { NULL, (uint)-1, 0 };
 
 //******************************************************************************
 // statics
